feat(TDP-3190): Add usage monitoring to cache
[Monitoring]
* Allow monitor of In/OutputStream when used as return of methods.

[Cache]
* Add timing and volume monitoring to cache service operations.
